backend finished

+++ backend/app/Http/Controllers/BookController.php
@@ -4,9 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\BookCollection;
 use App\Http\Resources\BookResource;
-use App\Models\User;
 use App\Models\Book;
+use App\Models\Image;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class BookController extends Controller
@@ -14,7 +15,8 @@ class BookController extends Controller
     public function index()
     {
         $books = Book::all();
-        return new BookCollection($books);
+        // return new BookCollection($books);
+        return BookResource::collection($books);
     }
 
     public function store(Request $request)
@@ -23,7 +25,8 @@ class BookController extends Controller
             'title' => 'required|string|max:255',
             'author' => 'required|string|max:255',
             'description' => 'required|string|max:255',
-            'price' => 'required|integer'
+            'price' => 'required|numeric',
+            'image' => 'image'
         ]);
 
         if ($validator->fails()) {
@@ -37,45 +40,61 @@ class BookController extends Controller
             'price' => $request->price,
         ]);
 
-        return response()->json(['Book created successfully', new BookResource($book)]);
+        if ($request->hasFile('image')) {
+            if (is_null($request->file('image'))) {
+                return response()->json(['message' => 'book created successfully', 'book' => new BookResource($book)], 200);
+            }
+            $path = $request->file('image')->store('/public/images');
+            $url = Storage::url($path);
+            $image = Image::create(['url' => $url, 'path' => $path]);
+            $book->imageid = $image->id;
+            $book->save();
+        } else {
+            return response()->json(['message' => 'book created successfully', 'book' => new BookResource($book)], 200);
+        }
+        return response()->json(['message' => 'book created successfully', 'book' => new BookResource($book)], 200);
+
+        // return response()->json(['message' => 'book created successfully', 'book' => new BookResource($book)], 200);
     }
 
-    public function show(Request $request)
+    public function show($author)
     {
-        $books = Book::where('author', $request->author)->get();
+        $books = Book::get()->where('author', $author);
 
         if ($books->isEmpty()) {
-            return response()->json('Data not found', 404);
+            return response()->json(['message' => 'book not found'], 403);
         }
-        return new BookCollection($books);
+        return BookResource::collection($books);
     }
 
-    public function update(Request $request, Book $book)
+    public function update(Request $request, $bookid)
     {
         $validator = Validator::make($request->all(), [
-            'title' => 'required|string|max:255',
-            'author' => 'required|string|max:255',
-            'description' => 'required|string|max:255',
-            'price' => 'required|integer'
+            'price' => 'required|numeric'
         ]);
 
         if ($validator->fails()) {
             return response()->json($validator->errors());
         }
 
-        $book->title = $request->title;
-        $book->author = $request->author;
-        $book->description = $request->description;
-        $book->price = $request->price;
+        $book = Book::find($bookid);
 
-        $book->save();
+        if (is_null($book)) {
+            return response()->json(['message' => 'book not found'], 403);
+        }
 
-        return response()->json(['Book updated successfully', new BookResource($book)]);
+        $book->update(['price' => $request->price]);
+
+        return response()->json(['message' => 'book updated successfully', 'book' => new BookResource($book)], 200);
     }
 
-    public function destroy(Book $book)
+    public function destroy($bookid)
     {
+        $book = Book::find($bookid);
+        if (is_null($book)) {
+            return response()->json(['message' => 'book not found'], 403);
+        }
         $book->delete();
-        return response()->json('Book deleted successfully');
+        return response()->json(['mesage' => 'book deleted successfully'], 200);
     }
 }
